move Action to logical plan file
