+++ rust/ballista/src/physical_plan.rs
@@ -18,6 +18,7 @@ use crate::arrow::datatypes::Schema;
 use crate::arrow::record_batch::RecordBatch;
 use crate::datafusion::logicalplan::LogicalPlan;
 
+///
 #[derive(Debug, Clone)]
 pub enum Action {
     Collect { plan: LogicalPlan },
