+++ rust/ballista/src/serde/from_proto.rs
@@ -21,7 +21,7 @@ use crate::datafusion::logicalplan::{
 };
 
 use crate::error::{ballista_error, BallistaError};
-use crate::plan::Action;
+use crate::physical_plan::Action;
 use crate::protobuf;
 
 impl TryInto<LogicalPlan> for protobuf::LogicalPlanNode {
